refactor: extract bucket card actions and meta into separate components
